flashb: Allow I2C commands to be customised with the config file
flashb.config contains a description of all of the configuration
settings.  Values for the motor board and jointio board are included.

This patch also adds command line option support.  These options allow
the selection of the board to program and the ELF file to load into
the board.

flashb -n motor face

Will load the motor board from the ELF file "face".

git-svn-id: http://svn.srobo.org/slug/utils/flashb/trunk@2080 98f67ba8-b320-0410-bda0-b753e294843c