flashb: Add a "give-up" option to allow for devices that aren't present.
When the robot boots, we don't know which peripherals are present.
flashb therefore needs to give up trying to update the firmware after
a number of attempts to talk to the device.

Please note that this is different to giving up on talking to a device
from the user's Python.  The users know which devices their code
requires to present.

git-svn-id: http://svn.srobo.org/slug/utils/flashb/trunk@2424 98f67ba8-b320-0410-bda0-b753e294843c